Implement the first user story in the API

+++ lib/github_interactor.rb
@@ -1,11 +1,12 @@
 # frozen_string_literal: true
 
 module GithubInteractor
-  def self.fetch_starred(username)
+  def self.fetch_starred(username, reply = :json)
     url = "https://api.github.com/users/#{username}/starred?sort=updated&direction=desc"
     starred_repos = RestClient.get(url)
-    JSON.parse(starred_repos, symbolize_names: true)
+    repos = JSON.parse(starred_repos, symbolize_names: true)
       .map { |repo| repo.slice(*%i[id name description html_url language]) }
-      .each { |repo| repo[:github_id] = repo.delete(:id) }
+    repos.each { |repo| repo[:github_id] = repo.delete(:id) } if reply == :html
+    repos
   end
 end
